Cria componente com comportamente de copiar o token
